+++ tests/Pub/TestClasses/Events/UserRetrievedWithCustomPayload.php
@@ -1,20 +1,18 @@
 <?php
 
-namespace PodPoint\AwsPubSub\Tests\Pub\Dummies\Events;
+namespace PodPoint\AwsPubSub\Tests\Pub\TestClasses\Events;
 
 use Illuminate\Broadcasting\InteractsWithSockets;
 use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
-use PodPoint\AwsPubSub\Tests\Pub\Dummies\Models\User;
+use PodPoint\AwsPubSub\Tests\Pub\TestClasses\Models\User;
 
-class UserRetrievedWithPublicProperties implements ShouldBroadcast
+class UserRetrievedWithCustomPayload implements ShouldBroadcast
 {
     use Dispatchable, InteractsWithSockets, SerializesModels;
 
     public $action = 'retrieved';
-
-    public $foo = 'bar';
 
     public $user;
 
@@ -26,5 +24,16 @@ class UserRetrievedWithPublicProperties implements ShouldBroadcast
     public function broadcastOn()
     {
         return ['users'];
+    }
+
+    public function broadcastWith()
+    {
+        return [
+            'action' => $this->action,
+            'data' => [
+                'user' => $this->user,
+                'foo' => 'baz',
+            ],
+        ];
     }
 }
