+++ tests/Pub/TestClasses/Models/UserWithBroadcastingEventsWithCustomPayloadForSpecificEvents.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace PodPoint\AwsPubSub\Tests\Pub\Dummies\Models;
+namespace PodPoint\AwsPubSub\Tests\Pub\TestClasses\Models;
 
 use PodPoint\AwsPubSub\Pub\Database\Eloquent\BroadcastsEvents;
 
-class UserWithBroadcastingEventsWithCustomPayload extends User
+class UserWithBroadcastingEventsWithCustomPayloadForSpecificEvents extends User
 {
     use BroadcastsEvents;
 
@@ -13,13 +13,18 @@ class UserWithBroadcastingEventsWithCustomPayload extends User
         return ['users'];
     }
 
+    public function broadcastEvents()
+    {
+        return ['updated'];
+    }
+
     public function broadcastWith($event)
     {
         return [
             'action' => $event,
             'data' => [
                 'user' => $this,
-                'foo' => 'bar',
+                'foo' => 'baz',
             ],
         ];
     }
