+++ tests/Sub/SqsSnsJobTest.php
@@ -17,6 +17,8 @@ class SqsSnsJobTest extends TestCase
 
     protected function setUp(): void
     {
+        parent::setUp();
+
         $this->sqsClient = $this->getMockBuilder(SqsClient::class)
             ->disableOriginalConstructor()
             ->getMock();
